requirements-ubuntu.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
et-xmlfile==1.1.0
Flask==3.0.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
lxml==4.9.3
MarkupSafe==2.1.3
multitasking==0.0.11
numpy==1.26.1
openpyxl==3.1.2
panda==0.3.1
pandas==2.1.2
passlib==1.7.4
peewee==3.17.0
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1
Werkzeug==3.0.1
yfinance==0.2.31
